Revert commit "Prepare for packaging internally" 44b6dd662adfbf0b56bb8079e1ef34ffdd293c4c
