+++ plugin-build/plugin/build.gradle.kts
@@ -50,18 +50,6 @@ gradlePlugin {
      vcsUrl = PluginBundle.VCS
 }
 
-publishing {
-    repositories {
-        maven {
-            setUrl("https://maven.pkg.github.com/doublesymmetry/kotlin-native-spm")
-            credentials {
-                username = System.getenv("GITHUB_USER")
-                password = System.getenv("GITHUB_TOKEN")
-            }
-        }
-    }
-}
-
 tasks.create("setupPluginUploadFromEnvironment") {
     doLast {
         val key = System.getenv("GRADLE_PUBLISH_KEY")
