--- plugin-build/plugin/src/main/java/com/doublesymmetry/kotlin/native/spm/entity/DslMarkers.kt
@@ -1,19 +1,0 @@
-package com.doublesymmetry.kotlin.native.spm.entity
-
-@DslMarker
-annotation class DependencyMarker
-
-@DslMarker
-annotation class TargetMarker
-
-@DslMarker
-annotation class TargetDependencyMarker
-
-@DslMarker
-annotation class PlatformMarker
-
-@DslMarker
-annotation class ProductMarker
-
-@DslMarker
-annotation class SupportedProductMarker
