--- plugin-build/plugin/src/main/java/com/doublesymmetry/kotlin/native/spm/entity/Entities.kt
@@ -1,9 +1,0 @@
-package com.doublesymmetry.kotlin.native.spm.entity
-
-import org.gradle.api.Named
-
-interface Product : Named
-
-interface Target : Named
-
-interface TargetDependency : Named
