--- plugin-build/plugin/src/main/java/com/doublesymmetry/kotlin/native/spm/entity/impl/TargetDependencyManager.kt
@@ -1,34 +1,0 @@
-package com.doublesymmetry.kotlin.native.spm.entity.impl
-
-import com.doublesymmetry.kotlin.native.spm.entity.TargetDependency
-import com.doublesymmetry.kotlin.native.spm.entity.TargetDependencyMarker
-
-@TargetDependencyMarker
-class TargetDependencyManager {
-    val targetDependencies = mutableListOf<TargetDependency>()
-
-    fun target(name: String, condition: String? = null) {
-        val target = Target(name, condition)
-        targetDependencies.add(target)
-    }
-
-    fun product(name: String, dependency: String, condition: String? = null) {
-        val product = Product(name, dependency, condition)
-        targetDependencies.add(product)
-    }
-
-    data class Target(
-        private val name: String,
-        val condition: String? = null
-    ) : TargetDependency {
-        override fun getName(): String = name
-    }
-
-    data class Product(
-        private val name: String,
-        val dependency: String,
-        val condition: String? = null
-    ) : TargetDependency {
-        override fun getName(): String = name
-    }
-}
