--- plugin-build/plugin/src/main/java/com/doublesymmetry/kotlin/native/spm/tasks/CleanSwiftPackageProjectTask.kt
@@ -1,24 +1,0 @@
-package com.doublesymmetry.kotlin.native.spm.tasks
-
-import com.doublesymmetry.kotlin.native.spm.plugin.KotlinSpmPlugin
-import com.doublesymmetry.kotlin.native.spm.swiftPackageBuildDirs
-import org.gradle.api.DefaultTask
-import org.gradle.api.provider.Property
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.TaskAction
-import org.jetbrains.kotlin.konan.target.Family
-
-abstract class CleanSwiftPackageProjectTask : DefaultTask() {
-    init {
-        description = "Clean swift package project in build directory"
-        group = KotlinSpmPlugin.TASK_GROUP
-    }
-
-    @Input
-    val platformFamily: Property<Family> = project.objects.property(Family::class.java)
-
-    @TaskAction
-    fun action() {
-        project.swiftPackageBuildDirs.platformRoot(platformFamily.get()).deleteRecursively()
-    }
-}
