--- plugin-build/plugin/src/main/java/com/doublesymmetry/kotlin/native/spm/utils/StringUnaryPlusContainer.kt
@@ -1,10 +1,0 @@
-package com.doublesymmetry.kotlin.native.spm.utils
-
-@Suppress("UnnecessaryAbstractClass")
-abstract class StringUnaryPlusContainer {
-    val container = mutableListOf<String>()
-
-    operator fun String.unaryPlus() {
-        container.add(this)
-    }
-}
