+++ plugin-build/plugin/src/main/java/com/github/pagr0m/kotlin/native/spm/entity/Entities.kt
@@ -1,0 +1,9 @@
+package com.github.pagr0m.kotlin.native.spm.entity
+
+import org.gradle.api.Named
+
+interface Product : Named
+
+interface Target : Named
+
+interface TargetDependency : Named
